Sensormodul klar?

+++ Sensormodul/Sensormodul/gyro.h
@@ -10,7 +10,6 @@
 #ifndef GYRO_H_
 #define GYRO_H_
 
-//Tveksamt om detta ska vara kvar
 #define ACTIVATE_ADC 0b10010100
 #define START_CONVERSION 0b10010100
 #define POLL 0b10000000
@@ -19,8 +18,8 @@
 void activate_adc();
 void start_conversion();
 uint16_t get_angular_rate();
-int adc_to_angular_rate(uint16_t data);
-int rotate_to(int angle);
+double adc_to_angular_rate(uint16_t data);
+double rotate_to(int angle);
 int get_is_rotated();
 void deactivate_adc();
 void has_rotated(int bit);
